test: hoist peers
